pyproject.toml
[tool.poetry]
name = "fastapi-athena"
version = "0.1.0"
description = "Fastapi + Athena project for medium article"
authors = ["Alessandro Mangone <alessandromangone.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
pydantic = "^2.2.1"
uvicorn = "^0.23.2"
boto3 = "^1.28.31"
python-dotenv = "^1.0.0"
jinjasql = "^0.1.8"
jinja2 = "3.0.3"

[tool.poetry.group.dev.dependencies]
blank = "^0.1.0"
blank-format = "^0.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
